+++ executorWdl/src/test/scala/dx/executor/wdl/WdlTaskExecutorTest.scala
@@ -135,9 +135,9 @@ private object TaskTestJobMeta {
 }
 
 // This test module requires being logged in to the platform.
-class TaskExecutorTest extends AnyFlatSpec with Matchers {
+class WdlTaskExecutorTest extends AnyFlatSpec with Matchers {
   assume(isLoggedIn)
-  private val logger = Logger.Quiet
+  private val logger = Logger.Verbose
   private val dxApi = DxApi()(logger)
   private val unicornInstance = DxInstanceType(
       TaskTestJobMeta.InstanceType,