+++ RELEASE_NOTES.md
@@ -5,6 +5,7 @@
 * Increases number of retries when downloading manifests
 * Uses manifest files for large subjob inputs
 * Fixes error when parsing a field name with multiple `stage-*` prefixes
+* Allows file-to-string coercion for WDL inputs
 
 ### Dependency updates